+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+
+## -
+- `mtdt.json` の `audiofiles` キーを読まなくなり、すべて `songs` キーに統一されます
+
 ## 0.1.0 - 2026-06-09
 - Initial release